* Import submodules from package init * Move DATA_PATH and NUMPY_DTYPE to utils so that there is no circular dependency
