+++ art/attacks/elastic_net.py
@@ -27,7 +27,7 @@ import logging
 import numpy as np
 import six
 
-from art import NUMPY_DTYPE
+from art.utils import NUMPY_DTYPE
 from art.classifiers.classifier import ClassifierGradients
 from art.attacks.attack import Attack
 from art.utils import compute_success, get_labels_np_array, check_and_transform_label_format
