+++ art/attacks/fast_gradient.py
@@ -27,7 +27,7 @@ import logging
 
 import numpy as np
 
-from art import NUMPY_DTYPE
+from art.utils import NUMPY_DTYPE
 from art.classifiers.classifier import ClassifierGradients
 from art.attacks.attack import Attack
 from art.utils import compute_success, get_labels_np_array, random_sphere, projection, check_and_transform_label_format
