+++ art/attacks/hop_skip_jump.py
@@ -27,7 +27,7 @@ import logging
 
 import numpy as np
 
-from art import NUMPY_DTYPE
+from art.utils import NUMPY_DTYPE
 from art.attacks.attack import Attack
 from art.utils import compute_success, to_categorical, check_and_transform_label_format
