+++ art/attacks/projected_gradient_descent.py
@@ -29,7 +29,7 @@ import logging
 
 import numpy as np
 
-from art import NUMPY_DTYPE
+from art.utils import NUMPY_DTYPE
 from art.classifiers.classifier import ClassifierGradients
 from art.attacks.fast_gradient import FastGradientMethod
 from art.utils import compute_success, get_labels_np_array, check_and_transform_label_format
