+++ art/attacks/virtual_adversarial.py
@@ -26,7 +26,7 @@ import logging
 
 import numpy as np
 
-from art import NUMPY_DTYPE
+from art.utils import NUMPY_DTYPE
 from art.classifiers.classifier import ClassifierNeuralNetwork, ClassifierGradients
 from art.attacks.attack import Attack
