+++ art/classifiers/tensorflow.py
@@ -478,7 +478,7 @@ class TensorFlowClassifier(ClassifierNeuralNetwork, ClassifierGradients, Classif
         from tensorflow.python.saved_model.signature_def_utils_impl import predict_signature_def
 
         if path is None:
-            from art import DATA_PATH
+            from art.utils import DATA_PATH
             full_path = os.path.join(DATA_PATH, filename)
         else:
             full_path = os.path.join(path, filename)
@@ -556,7 +556,7 @@ class TensorFlowClassifier(ClassifierNeuralNetwork, ClassifierGradients, Classif
             import tensorflow.compat.v1 as tf
             tf.disable_eager_execution()
         from tensorflow.python.saved_model import tag_constants
-        from art import DATA_PATH
+        from art.utils import DATA_PATH
 
         full_path = os.path.join(DATA_PATH, state['model_name'])
